Input trace decay is a learnable parameter now as well.
Pytorch by default doesn't like accessing the .grad attribute of non leaf tensors, but splitting the graph into two is mathematically identical. The average input is calculated as i=t(1-d), so we can rerun the exact same calculation a second time and .backward(ls) on it because of the chain rule it makes sense, the derivative of the loss w.r.t. the average input is the same regardless of if the input trace decay is part of the graph or not. Then we do one optimizer.step() to update all the learnable parameters in one go.