Estimating firing frequency from input, decay and threshold.
I think I got it to work? Effectively, we can plot the membrane level over time, it's (i-id^x)/(1-d). We want to find out how long it takes for the membrane level to reach the threshold, so we take the inverse of the function, and it's the whole log formula. We can do a simple check if the neuron fires at all by finding out if the threshold is below the stabilization level i/(1-d), at which point we construct a piecewise function to plot the thing. In the valid, firing range, we use the proper log formula, otherwise we use a rapidly accelerating exponential formula and take the reciprocal to get a small number. It's not exact, but it's differentiable, and also, close enough. After all, it's more of an indication that the firing frequency should be around zero, not actually 1e-2 or whatever. I hope it works.