LIF surrogate derivative
Updated so that now the excess function is made to be 0 at the midpoint between t=i and t=i/(1-d). This way, when applying sigmoid and we obtain the frequency, it will be 0.5 at the midpoint, contrary to before where it was at t=i/(1-d). It differs from my previous formula (the garbage one) because I don't scale the sigmoid function to dilute it. Possibly that was the issue beforehand? Or perhaps just a garbage implementation. In any case, it works, up next to do an actual test on MNIST or something, although I'd like to make a softmax neuron as well.