LIF testing
It works. List shuffling, as well as the loss on only the last output, works to make the layers memorize various patterns. I also messed up and accidentally originally made it 10>10>40, while it should've been 10>40>10, so that's also fixed. Have not played around with learning rate yet, but it seems that 1e-3 works, I'd assume 1e-2 would too.