Classification testing
Easily memorizes random input to predictable output. The test targets each of the output neurons with random inputs. Since it can memorize white noise to output mapping, then that means that in theory, when we have a perfect, maximally compressing encoder, we can decode the output. Maximal compression is indistinguishable from noise, so at least the "decoder" part definitely works.