Generalization Testing
Changed the raw input MNIST image to instead be rate coded via bernoulli sampling. Changing the alpha would change the maximum brightness, the lower it is the less likely the input neuron is to fire, and thus straining the model. It shouldn't improve accuracy or loss, it's more so a test on if the model can generalize or work with messy data.