Poisson Click Test
The poisson click test is made to see if the model can count and store information over a long time. The concept is that it listens to the input, which will be a signal of noise on the left or the right, then it should internally add up which side had more noise, and then at the end, when it sees the recall signal, it should classify into which side has had more noise. Vibecoded the poisson generator