Testing hyperparameters
Added a default argument for scaling of the weights. They are initialized as randn, although I should probably fix that I guess, but honestly it seems decent enough. I mean the vast majority of weights are weak, for fine control, a small amount are for big influences, what's the issue? Anyways, point being, I've played with different stuff, like changin the amount of thinking steps, the architecture and the number of layers, and no matter what the model seems to stabilize around 93% accuracy? Admittedly, the inputs are noisy as hell, the actual number being only 0.3 brightness while the rest is at 0.1, but that seems... good? I mean decent. Training is long as hell. If I had this stuff sorted as a library I'd want to play with ipynb files to plot the weights and stuff, see how each layer monitors stuff, like time or whatnot, like for example the input decay, membrane decay, the likes. Although thats a mini experiment in itself. I'm next gonna make a manager for all the losses and accuraccies and their running averages, ts pmo rn.