Added rendering of input trace decay of layer
Added a function to render tensors as images, but the main goal is to plot the states and parameters of layers. Pretty cool, I want to check out if the layers are scale invariant or not by looking at the distributions of values between layers and comparing them, will make a function for that next.