Created testing folder for testing out various things in the works.
Created _reinforce_manager.py for the trace alternative to the REINFORCE algorithm, the idea is to instead look back into the past and reflect on it (perhaps rename to REFLECT?). Also created entropy_testing.py for the idea of comparing the input received to the expected input based on the average_input calculated from the trace, concept being that if the entropy is consistently high, then the decay should be decreased since this implies rapidly changing inputs and hence it possibly suggests to instead make this neuron more rapid response? On the contrary predictable results means that it's working long term stuff? idk, that's why this stuff is in testing/