Completed refactoring to nn.Module
nn modules are now called snn as not to confuse with pytorch.nn; all snn modules are now built around nn.Module, allowing for models to be easily moved around, saved, and managed by one optimizer rather than the many local ones. The use of nn.Parameter allows for models to be easily cleared of gradients to None, regardless of what is and isn't added to the optimizer.