Solved weight initialization for scale invariance
The idea is that with initialization, the signal should be able to propagate regardless of how many neurons are present in the previous layer, or the number of layers. Effectively, given that some fraction of the input neurons fired, the same fraction of output neurons should fire too. The idea is that hence if 100% of the input neurons fire, then each output neuron has a 100% chance too, a 50% firing frequency should correlate to a 50% chance, etc. Some weights should be negative for inhibition. For LIF, the mean should be t/i so that i samples = t, and the stdev should be the sqrt(t/i) so that i samples = t. With this setup, the notion is preserved, because if the number of samples is i/2, the mean is t/2, etc. For LIS the mean is 0 and the stdev is sqrt(1/i) so that i samples = 1. This way, extra neurons in each layer dictate the precision and ability of fine control, rather than having a big range.