Fixed REFLECT learning signal calculation
In normal RL, the learning signal is the derivative of the loss w.r.t. the probability, given that the loss is `-advantage * sum(choice * log(probability))`, it's quite simple since the entire computation graph already has everything to get `d_L/d_probability`, but in the case of REFLECT, we are storing the trace of the logprobs, and the raw probabilities themselves aren't used in the computation graph. Therefore the best we can return is average_logprob.grad, but LIS expected `d_L/d_probability`. Subsequently, the solution is to just adjust LIS so that it _can_ expect `d_L/d_logprob`, since applying log to the probability inside the LIS .backward() is easy as, and then on that is called .backward(-ls). The learning signal is negative because log returns the opposite of what you want, a signal saying "decrease" really means the opposite.