Fixing the .backward() gradient approximation for LIF
Last time it was done with sigmoid and whatnot. For some reason, no matter what I do the accuracy is never over 90% for MNIST, which is strange. I have an idea that the gradient approximation is wrong, which makes sense. Visually playin around with the real formulas is nothing like what the sigmoid approximation does, so now I want to fix it.