Generalized hidden state management
Beforehand, you'd have to manually call the helper methods, or alternatively use the Sequential equivalent to manage the states in the sequential module, but you'd still have to manually manage it for more complex architectures. Now, there's an `snn.TTModule` which is an `nn.Module` child, and you can just call that on any class instance that's a child of TTModule and you'll get all the states cleared reset or whatever. It internally does recursive checks to make sure that all attributes and attribute's attributes are cleared and whatnot. Point being, make any model you like as before, just make it a `snn.TTModule` child instead of `nn.Module` and you're good to go on clearing that exact instance.