Refactored `view_tuple` logic into `dim`; refactored `*_is_vector` logic into `*_rank` logic; removed parameter decomposition logic.
A lot simpler now. Instead of the view tuple needing to be whatever view of the parameter you need so that the parameter shape is adjusted to the tensor shape, it's now based on `dim`, where that dimension is briefly moved to the last dimension where the parameter can hence be easily applied, and then moved back. Unlike the old approach where you would also have to consider batches, here you can just say -3 for any cnn, regardless of if it's batched or not, also looks more pytorchian. The `*_is_vector` would get real ugly real quick with recurrent layers that have a matrix instead of a scalar or vector, so it's easier to just work with ranks directly. Also got rid of the parameter decomposition since it's wasting space, and would waste all the more space when it comes to the matrix based recurrent layers, and hence I'd just move to if based logic and then at that point might as well just not do the decomposition in the first place.